requirements.txt
# requirements.txt

# Web 框架
fastapi>=0.100.0
uvicorn>=0.23.0

# 数据处理
pydantic>=2.0.0
pyyaml>=6.0

# HTTP 客户端
httpx>=0.24.0

# 测试
pytest>=7.0.0
pytest-cov>=4.0.0

# E2E 测试（可选）
playwright>=1.40.0
pytest-playwright>=0.4.0
